+++ crates/rustc_codegen_gpu/tests/codegen/subslice.rs
@@ -9,7 +9,7 @@ pub fn kernel_arith(a: &[u8], b: &mut u8) {
 }
 
 #[gpu::kernel]
-pub fn kernel_arith_wrapper(a: &[u8], a_window: usize, b: &mut [u8], b_window: u32) {
+pub fn kernel_arith_wrapper(a: &[u8], a_window: usize, b: &mut [u8], b_window: usize) {
     let mut b_local = gpu::chunk_mut(b, gpu::MapLinear::new(b_window));
 
     kernel_arith(a, &mut b_local[0]);
